requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.11.12
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.121.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.44.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.44.0
tomli==2.3.0
typer==0.20.0
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
